Overwrite state dict loading at class level
This way, loading arbitrary PyTorch models should be simplified.
This mostly concerns loading pretrained PyTorch models.
We'd like to make this modification in a initialization context.
